Canonical URLs — Prevent duplicate indexing using Next.js Metadata API
