+++ frontend/src/app/layout.tsx
@@ -8,7 +8,6 @@ import Footer from "./components/Footer";
 import Navbar from "./components/Navbar";
 import ScrollToTop from "./components/ScrollToTop";
 import { ThemeProvider } from "./providers";
-import CanonicalURL from "./components/CanonicalURL";
 
 const inter = Inter({
 	subsets: ["latin"],
@@ -40,9 +39,7 @@ export const metadata: Metadata = {
 	title: "EnvForage | ML Environment Provisioning",
 	description:
 		"Generate intelligent, safe, and deterministic ML/AI environment setup scripts.",
-	// NOTE: Per-page canonical URLs are set via individual page metadata exports
-	// and the <CanonicalURL /> client component mounted below in <head>.
-	// Do NOT set a root-level canonical here — it would override every page with "/".
+	// Per-page canonical URLs are set via individual page metadata exports.
 };
 
 export default function RootLayout({
@@ -53,9 +50,6 @@ export default function RootLayout({
 	return (
 		<html lang="en" suppressHydrationWarning>
 			<head>
-				{/* Canonical URL — prevents duplicate indexing across trailing-slash,
-				    query-string, and www/non-www variants for every route. */}
-				<CanonicalURL />
 				<Script id="theme-init" strategy="beforeInteractive">
 					{`
             try {
